Fix GetWorkspaceName for GitHub Codespaces.
GitHub Codespaces sticks a few "dot folders" in the /workspaces root. These folders cause GetWorkspaceName to return multiple results, breaking downstream functions. This change adds a filter to the find command to exclude folders beginning with a period.

Alternatively, this could be modified to check for and use the CODESPACE_VSCODE_FOLDER environment variable if available.